mv rest of accessors_metadata to core
